requirements.txt
# 网络热门事件抓取与可视化系统 - Python 依赖清单
# 安装命令: pip install -r requirements.txt

# Web 框架
Flask==3.0.2
Werkzeug==3.0.1

# 数据库
PyMySQL==1.1.0
SQLAlchemy==2.0.28

# 数据处理
pandas==2.0.3
numpy==1.24.4
pytz==2024.1
tzdata==2024.1

# 网络请求
requests==2.31.0
urllib3==2.2.1
charset-normalizer==3.3.2
idna==3.6
certifi==2024.2.2

# 中文分词与情感分析
jieba==0.42.1
snownlp==0.12.3

# 可视化
matplotlib==3.7.5
wordcloud==1.9.3
Pillow==10.2.0

# 辅助工具
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
six==1.16.0
cycler==0.12.1
kiwisolver==1.4.5
contourpy==1.1.1
fonttools==4.50.0
pyparsing==3.1.2
packaging==24.0
setuptools==56.0.0
importlib-metadata==7.0.2
importlib-resources==6.3.1
zipp==3.18.1
colorama==0.4.6
typing-extensions==4.10.0
greenlet==3.0.3
